requirements.txt
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Flask==2.2.2
flask-redis==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
Werkzeug==2.2.2
wrapt==1.14.1
